Introduced CryptoUtils. Renamed DetRandomProvider to avoid confusion.
