+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/system/address/Address.java
@@ -4,7 +4,6 @@ package com.swirlds.platform.system.address;
 import static com.swirlds.common.utility.NonCryptographicHashing.hash32;
 
 import com.swirlds.base.utility.ToStringBuilder;
-import com.swirlds.platform.crypto.CryptoStatic;
 import com.swirlds.platform.crypto.SerializableX509Certificate;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
@@ -14,6 +13,7 @@ import java.security.cert.CertificateEncodingException;
 import java.security.cert.X509Certificate;
 import java.util.Arrays;
 import java.util.Objects;
+import org.hiero.base.crypto.CryptoUtils;
 import org.hiero.base.crypto.SerializablePublicKey;
 import org.hiero.base.io.SelfSerializable;
 import org.hiero.base.io.streams.SerializableDataInputStream;
@@ -641,7 +641,7 @@ public class Address implements SelfSerializable {
         if (certificate == null) {
             return null;
         }
-        return CryptoStatic.checkCertificate(certificate.getCertificate()) ? certificate : null;
+        return CryptoUtils.checkCertificate(certificate.getCertificate()) ? certificate : null;
     }
 
     /**
