+++ platform-sdk/swirlds-platform-core/src/test/java/com/swirlds/platform/crypto/SerializableX509CertificateTests.java
@@ -19,7 +19,7 @@ import java.security.SignatureException;
 import java.security.cert.CertificateEncodingException;
 import java.security.cert.X509Certificate;
 import java.util.Random;
-import org.hiero.base.crypto.internal.CryptoUtils;
+import org.hiero.base.crypto.internal.DetRandomProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -37,7 +37,7 @@ class SerializableX509CertificateTests {
         final int ecKeySize = 384;
 
         final Random nonSecureRandom = getRandomPrintSeed();
-        final SecureRandom secureRandom = CryptoUtils.getDetRandom();
+        final SecureRandom secureRandom = DetRandomProvider.getDetRandom();
         secureRandom.setSeed(nonSecureRandom.nextLong());
 
         // Render key pairs.
